+++ back/src/bsda/permissions.ts
@@ -1,0 +1,41 @@
+import { User, Bsda } from "@prisma/client";
+import { NotFormContributor } from "../forms/errors";
+import { getFullUser } from "../users/database";
+
+export async function checkIsFormContributor(
+  user: User,
+  form: Partial<
+    Pick<
+      Bsda,
+      | "emitterCompanySiret"
+      | "destinationCompanySiret"
+      | "transporterCompanySiret"
+      | "workerCompanySiret"
+    >
+  >,
+  errorMsg: string
+) {
+  const isContributor = await isFormContributor(user, form);
+
+  if (!isContributor) {
+    throw new NotFormContributor(errorMsg);
+  }
+
+  return true;
+}
+
+export async function isFormContributor(user: User, form: Partial<Bsda>) {
+  const fullUser = await getFullUser(user);
+  const userSirets = fullUser.companies.map(c => c.siret);
+
+  const formSirets = [
+    form.emitterCompanySiret,
+    form.destinationCompanySiret,
+    form.transporterCompanySiret,
+    form.workerCompanySiret
+  ];
+
+  const siretsInCommon = userSirets.filter(siret => formSirets.includes(siret));
+
+  return siretsInCommon.length;
+}