+++ back/src/bsda/resolvers/mutations/duplicate.ts
@@ -1,15 +1,15 @@
-import { Bsvhu, BsvhuStatus } from "@prisma/client";
+import { Bsda, BsdaStatus } from "@prisma/client";
 import { checkIsAuthenticated } from "../../../common/permissions";
 import getReadableId, { ReadableIdPrefix } from "../../../forms/readableId";
-import { MutationDuplicateBsvhuArgs } from "../../../generated/graphql/types";
+import { MutationDuplicateBsdaArgs } from "../../../generated/graphql/types";
 import prisma from "../../../prisma";
-import { expandVhuFormFromDb } from "../../converter";
+import { expandBsdaFormFromDb } from "../../converter";
 import { getFormOrFormNotFound } from "../../database";
 import { checkIsFormContributor } from "../../permissions";
-import { indexBsvhu } from "../../elastic";
+
 export default async function duplicate(
   _,
-  { id }: MutationDuplicateBsvhuArgs,
+  { id }: MutationDuplicateBsdaArgs,
   context
 ) {
   const user = checkIsAuthenticated(context);
@@ -24,8 +24,7 @@ export default async function duplicate(
 
   const newForm = await duplicateForm(prismaForm);
 
-  await indexBsvhu(newForm);
-  return expandVhuFormFromDb(newForm);
+  return expandBsdaFormFromDb(newForm);
 }
 
 function duplicateForm({
@@ -34,22 +33,28 @@ function duplicateForm({
   updatedAt,
   emitterEmissionSignatureAuthor,
   emitterEmissionSignatureDate,
+  workerWorkSignatureAuthor,
+  workerWorkSignatureDate,
   transporterTransportSignatureAuthor,
   transporterTransportSignatureDate,
-  destinationReceptionQuantityNumber,
-  destinationReceptionQuantityTons,
+  destinationReceptionQuantityType,
+  destinationReceptionQuantityValue,
+  destinationReceptionDate,
+  destinationReceptionSignatureAuthor,
+  destinationReceptionSignatureDate,
   destinationReceptionAcceptationStatus,
   destinationReceptionRefusalReason,
   destinationOperationCode,
   destinationOperationSignatureAuthor,
   destinationOperationSignatureDate,
+  destinationOperationDate,
   ...rest
-}: Bsvhu) {
-  return prisma.bsvhu.create({
+}: Bsda) {
+  return prisma.bsda.create({
     data: {
       ...rest,
       id: getReadableId(ReadableIdPrefix.VHU),
-      status: BsvhuStatus.INITIAL,
+      status: BsdaStatus.INITIAL,
       isDraft: true
     }
   });