+++ back/src/forms/readableId.ts
@@ -4,7 +4,8 @@ export enum ReadableIdPrefix {
   BSD = "BSD", // Bordereau de suivi des déchets dangereux "générique" (CERFA n° 12571*01)
   DASRI = "DASRI", // Bordereau de suivi des déchets d'activités de soins à risque infectieux
   FF = "FF", // Bordereau de suivi de déchets fluides frigorigènes
-  VHU = "VHU" // Bordereau de suivi des véhicules hors d'usage
+  VHU = "VHU", // Bordereau de suivi des véhicules hors d'usage
+  BSDA = "BSDA" // Bordereau de suivi des déchets d'amiante
 }
 
 /**