+++ Changelog.md
@@ -56,6 +56,7 @@ et le projet suit un schéma de versionning inspiré de [Calendar Versioning](ht
 - Ajout de différents filtres dans les tableaux [PR 810](https://github.com/MTES-MCT/trackdechets/pull/810)
 - Ajout d'un sélecteur de type de bordereaux dans les tableaux, en vue des prochains bordereaux [PR 810](https://github.com/MTES-MCT/trackdechets/pull/810)
 - Ajout de la possibilité de créer des bordereaux DASRI via l'API [PR 787](https://github.com/MTES-MCT/trackdechets/pull/787)
+- Ajout du bordereau de suivi des déchets d'aminate à l'API [PR 873](https://github.com/MTES-MCT/trackdechets/pull/873)
 
 #### :boom: Breaking changes